Move new migration and DB restore scripts into pgsql. Update README file to use the new DB restore script. Change daily backup script to make a custom backup (rather than a text file backup). These backups take much less space and can be restored many times faster than a text backup.

+++ pgsql/restoreRenamedDbFromBackup.sh
@@ -7,12 +7,27 @@ if [[ "$db" == "" || "$dump" == "" ]]; then
     exit 1
 fi
 
+# Script directory.
+SDIR=`dirname $0`
+
+# Make DB user ***REMOVED*** into a superuser temporarily.
+psql -f $SDIR/role_alter_su.sql -U postgres postgres
+if [[ $? != 0 ]]; then
+    exit 1
+fi
+
 psql -U ***REMOVED*** Africa <<EOD
 drop database "$db";
 create database "$db";
 EOD
 
 pg_restore -d "$db" -U ***REMOVED*** $dump
+
+# Revoke superuser role from DB user ***REMOVED***.
+psql -f $SDIR/role_normal.sql -U postgres postgres
+if [[ $? != 0 ]]; then
+    exit 1
+fi
 
 echo "Done! You will need to edit pg_hba.conf in the pgsql directory to support"
 echo "the new database name, copy it to the postgresql data directory, and restart"
